Rename panelStack2Example.tsx
