+++ packages/docs-app/src/examples/core-examples/panelStackExample.tsx
